backend/requirements.txt
fastapi>=0.110
uvicorn>=0.29
python-multipart>=0.0.9
pydantic>=2.6
aiofiles>=23.2
google-generativeai>=0.5.2
python-dotenv

# Text to speech
edge-tts>=6.1.10
